prueba3/Prueba3/requirements.txt
pillow
django-crispy-forms
cx_oracle
